tweak(tupaiaWeb): RN-1445: Update request access flow (#5901)
* Request project access modal in ui components

* Move datatrak web to use new ui component

* WIP

* Fix build issue

* WIP

* Remove old files from tupaia web

* Styling fixes

* Tidy ups

* PR fixes

* Update RequestProjectAccess.tsx

---------
