+++ packages/datatrak-web/src/api/queries/useCountryAccessList.ts
@@ -7,7 +7,7 @@ import { useQuery } from '@tanstack/react-query';
 import { Project, ProjectCountryAccessListRequest } from '@tupaia/types';
 import { get } from '../api';
 
-export const useCountryAccessList = (projectCode: Project['code']) => {
+export const useCountryAccessList = (projectCode?: Project['code']) => {
   return useQuery(
     ['me/countries', projectCode],
     (): Promise<ProjectCountryAccessListRequest.ResBody> =>
