+++ packages/datatrak-web/src/views/RequestProjectAccessPage.tsx
@@ -5,9 +5,10 @@
 import React from 'react';
 import styled from 'styled-components';
 import { Paper } from '@material-ui/core';
-import { RequestProjectAccess } from '../features';
 import { useNavigate, useSearchParams } from 'react-router-dom';
+import { RequestProjectAccess } from '@tupaia/ui-components';
 import { ROUTES } from '../constants';
+import { useCountryAccessList, useProject, useRequestProjectAccess } from '../api';
 
 const Wrapper = styled(Paper).attrs({
   variant: 'outlined',
@@ -21,13 +22,26 @@ export const RequestProjectAccessPage = () => {
   const navigate = useNavigate();
   const [searchParams] = useSearchParams();
   const projectCode = searchParams.get('project') || '';
+
+  const { data: project, isLoading: isLoadingProject, isFetched } = useProject(projectCode);
+  const { mutate: requestProjectAccess, isLoading, isSuccess } = useRequestProjectAccess();
+  const { data: countries } = useCountryAccessList(projectCode);
+
   const onReturnToProjects = () => {
     navigate(ROUTES.PROJECT_SELECT);
   };
 
   return (
     <Wrapper>
-      <RequestProjectAccess projectCode={projectCode} onClose={onReturnToProjects} />
+      <RequestProjectAccess
+        onClose={onReturnToProjects}
+        project={project}
+        onSubmit={requestProjectAccess}
+        isLoading={isLoadingProject || !isFetched}
+        isSubmitting={isLoading}
+        isSuccess={isSuccess}
+        countries={countries}
+      />
     </Wrapper>
   );
 };
