--- packages/tupaia-web/src/views/RequestProjectAccessModal/index.ts
@@ -1,6 +1,0 @@
-/*
- * Tupaia
- * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
- */
-
-export { RequestProjectAccessModal } from './RequestProjectAccessModal';
