+++ packages/ui-components/src/features/RequestProjectAccess/ProjectAccessForm.tsx
@@ -13,15 +13,15 @@ import {
   Typography,
 } from '@material-ui/core';
 import { CheckCircle } from '@material-ui/icons';
+import { ProjectCountryAccessListRequest, WebServerProjectRequest } from '@tupaia/types';
 import {
   Checkbox as BaseCheckbox,
-  Form,
-  FormInput,
   SpinningLoader,
   TextField,
-} from '@tupaia/ui-components';
-import { useCountryAccessList, useRequestProjectAccess } from '../../api';
-import { Button } from '../../components';
+  Button,
+  Alert,
+} from '../../components';
+import { Form, FormInput } from '../Form';
 
 const FormControl = styled(BaseFormControl).attrs({
   component: 'fieldset',
@@ -63,6 +63,9 @@ const TextArea = styled(TextField).attrs({
     font-size: 0.875rem;
     padding: 0.875rem;
   }
+  .MuiInputBase-root {
+    background-color: ${({ theme }) => theme.palette.background.paper};
+  }
   // we have to override this here as there are selectors inside ui-components with higher specificity than we can achieve via the theme overrides
   .MuiOutlinedInput-root.Mui-focused .MuiOutlinedInput-notchedOutline {
     box-shadow: none;
@@ -95,19 +98,38 @@ const LoaderWrapper = styled.div`
 /** Fixes janky spacing changes when 'Request access' button is enabled or disabled */
 const StyledDialogActions = styled(DialogActions)`
   gap: 1rem;
-
+  padding: 1.5rem 0 0 0;
   & > :not(:first-child) {
     margin-inline-start: 0;
   }
 `;
 
+const FormButton = styled(Button)`
+  text-transform: none;
+  font-size: 0.875rem;
+`;
+
+type Country = ProjectCountryAccessListRequest.ResBody[number];
+
 interface ProjectAccessFormProps {
-  project: any;
+  project?: WebServerProjectRequest.ResBody;
   onClose?: () => void;
+  countries: Country[];
+  onSubmit: (data: { entityIds: string[]; message: string; projectCode: string }) => void;
+  isSubmitting: boolean;
+  isSuccess: boolean;
+  closeButtonText?: string;
 }
 
-export const ProjectAccessForm = ({ project, onClose }: ProjectAccessFormProps) => {
-  const { data: countries } = useCountryAccessList(project.code);
+export const ProjectAccessForm = ({
+  project,
+  onClose,
+  countries,
+  onSubmit,
+  isSubmitting,
+  isSuccess,
+  closeButtonText = 'Back to projects',
+}: ProjectAccessFormProps) => {
   const formContext = useForm({
     mode: 'onChange',
   });
@@ -115,15 +137,37 @@ export const ProjectAccessForm = ({ project, onClose }: ProjectAccessFormProps) 
     formState: { isValid },
     register,
   } = formContext;
-  const { mutate: requestProjectAccess, isLoading, isSuccess } = useRequestProjectAccess();
 
   const projectCode = project?.code;
   const submitForm = (formData: any) => {
-    requestProjectAccess({
+    onSubmit({
       ...formData,
       projectCode,
     });
   };
+
+  // the countries that are available to request
+  const availableCountries = countries?.filter((c: Country) => !c.hasAccess && !c.hasPendingAccess);
+
+  // show the no countries message if the country access list has loaded and there are no countries available
+  const showNoCountriesMessage = !availableCountries?.length && !isSuccess;
+
+  if (showNoCountriesMessage) {
+    return (
+      <>
+        <Alert severity="info">
+          There are no countries available to request access to for this project. This means you
+          already have access to all countries in this project. If you need to change your
+          permissions, please contact your system administrator.
+        </Alert>
+        <StyledDialogActions>
+          <FormButton onClick={onClose} variant="text" color="default">
+            Back
+          </FormButton>
+        </StyledDialogActions>
+      </>
+    );
+  }
 
   // On success, show a success message to the user and direct them back to the projects list
   if (isSuccess)
@@ -141,36 +185,46 @@ export const ProjectAccessForm = ({ project, onClose }: ProjectAccessFormProps) 
         </SuccessWrapper>
 
         <DialogActions>
-          <Button onClick={onClose}>Back to Projects</Button>
+          <FormButton onClick={onClose}>{closeButtonText}</FormButton>
         </DialogActions>
       </>
     );
 
   // While the request is being processed, show a loading spinner
-  if (isLoading)
+  if (isSubmitting)
     return (
       <LoaderWrapper>
         <SpinningLoader />
       </LoaderWrapper>
     );
 
+  const getTooltip = (country: Country) => {
+    if (country.hasPendingAccess) {
+      return 'Approval in progress';
+    }
+    if (country.hasAccess) {
+      return 'You already have access to this country';
+    }
+    return undefined;
+  };
+
   return (
     <Form onSubmit={submitForm as SubmitHandler<any>} formContext={formContext}>
       <FormControl>
         <FormGroup>
-          {countries?.map(({ id, name, hasPendingAccess }) => {
+          {countries?.map(country => {
+            const { id, name, hasPendingAccess, hasAccess } = country;
+            const tooltip = getTooltip(country);
             return (
               <Checkbox
                 id={id}
                 inputRef={register({ validate: value => value.length > 0 })}
-                disabled={hasPendingAccess}
+                disabled={hasPendingAccess || hasAccess}
                 key={id}
                 label={name}
                 name="entityIds"
                 value={id}
-                tooltip={
-                  hasPendingAccess ? 'You have already requested access to this country' : undefined
-                }
+                tooltip={tooltip}
               />
             );
           })}
@@ -178,16 +232,16 @@ export const ProjectAccessForm = ({ project, onClose }: ProjectAccessFormProps) 
       </FormControl>
       <FormInput Input={TextArea} name="message" />
       <StyledDialogActions>
-        <Button variant="outlined" onClick={onClose}>
+        <FormButton variant="text" onClick={onClose} color="default">
           Back
-        </Button>
-        <Button
+        </FormButton>
+        <FormButton
           type="submit"
           disabled={!isValid}
           tooltip={isValid ? undefined : 'Select one or more countries to proceed'}
         >
           Request access
-        </Button>
+        </FormButton>
       </StyledDialogActions>
     </Form>
   );
