+++ packages/ui-components/src/features/RequestProjectAccess/index.ts
@@ -1,5 +1,5 @@
 /*
  * Tupaia
- *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ *  Copyright (c) 2017 - 2024 Beyond Essential Systems Pty Ltd
  */
 export { RequestProjectAccess } from './RequestProjectAccess';
