+++ packages/ui-components/src/features/index.ts
@@ -5,3 +5,4 @@
 export * from './Auth';
 export * from './Form';
 export { RouterLink } from './RouterLink';
+export * from './RequestProjectAccess';
